+++ src/main/java/insane96mcp/insanesurvivaloverhaul/data/generator/ISOItemTagsProvider.java
@@ -125,7 +125,7 @@ public class ISOItemTagsProvider extends ItemTagsProvider {
         tag(BigOreVeins.ORE_ROCKS_ITEM)
                 .add(BigOreVeins.COPPER_ORE_ROCK.item().get(), BigOreVeins.IRON_ORE_ROCK.item().get(), BigOreVeins.GOLD_ORE_ROCK.item().get());
 
-        tag(StackSizes.NO_STACK_SIZE_CHANGES)
+        tag(StackSizes.NO_FOOD_STACK_SIZE_CHANGES)
                 .add(Items.GOLDEN_CARROT);
     }